requirements.txt
# FastAPI and related
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
pydantic_settings==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
python_jose==3.3.0
sqlalchemy==1.4.23
alembic==1.7.1
langchain
langgraph==0.2.28
langchain_community
langchain_ollama
langchain_groq
openai==0.27.0
faiss-cpu==1.7.2
pytest==6.2.5
requests==2.26.0
httpx==0.18.2
python-dotenv==0.19.0
black==21.7b0
isort==5.9.3
flake8==3.9.2
numpy
pandas
logging
reportlab
json